requirements.txt
requests == 2.24.0
django == 2.2.12
djangorestframework==3.10.3
pygments == 2.6.1
